frontend/requirements-frontend.txt
pandas
requests
streamlit
python-multipart
uvicorn
Pillow